# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "guru-core API",
    "description": "\nThe single endpoint the app talks to. A user states a goal; the service collects\nwhatever context exists, asks follow-up questions when something essential is missing,\nand produces three difficulty variants of a scheduled plan.\n\n**Authentication.** Every endpoint except `POST /v1/auth/google`, `GET /health` and the\npresigned `/v1/files/*` routes needs `Authorization: Bearer <jwt>`. Get the token from\n`POST /v1/auth/google`; it carries the user id and expires after `JWT_TTL_SECONDS`.\nThe team-facing `role-models` writes authenticate with `X-API-Key` instead.\n\n**Long-running work.** Plan generation, import parsing and calendar export run on a\nqueue, so the endpoints that start them return `202` with an id. Poll the resource\n(`GET /v1/plan-sessions/{id}`, `GET /v1/imports`) rather than expecting a synchronous\nresult — PostgreSQL is the source of truth for job state, Redis only caches it.\n\n**Errors.** Every failure returns the same envelope:\n`{\"error\": {\"code\": \"not_found\", \"message\": \"...\"}}`. The code is the snake_case name of\nthe domain error: `invalid_input` (422), `unauthorized` (401), `forbidden` (403),\n`not_found` (404), `conflict` (409), `reauth_required` (409), `rate_limited` (429).\n",
    "contact": {
      "name": "Quasar-Gang",
      "url": "https://github.com/Quasar-Gang/guru-core"
    },
    "license": {
      "name": "Proprietary — all rights reserved"
...
  "paths": {
...
    "/v1/profile": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Read the signed-in user's profile",
        "description": "The standing context every plan is generated against: who the user is and where.\n\n`answers` holds the onboarding questionnaire as a free-form JSON object, and\n`timezone` is the IANA zone every scheduled task is placed in. The plan engine reads\nthis on each generation, so keeping it current directly improves plan quality.\n\nA user who has never written a profile is **not** an error: this returns\n`answers: {}`, `timezone: \"UTC\"` and the current time as `updated_at`. There is no\n`404` here — use `is_new_user` from login, or an empty `answers`, to decide whether to\nshow onboarding.",
        "operationId": "get_profile_v1_profile_get",
        "responses": {
          "200": {
            "description": "The stored answers and timezone, or the defaults when nothing has been saved yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileView"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized` — missing, invalid or expired bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` — too many requests for this user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerJWT": []
          }
        ]
      },
      "put": {
        "tags": [
          "profile"
        ],
        "summary": "Replace the signed-in user's profile",
        "description": "Write the onboarding answers, and later any edit the user makes to them.\n\nThis is a genuine `PUT`: `answers` **replaces** the stored object wholesale, it is not\nmerged key by key. To change one answer, read `GET /v1/profile`, edit the object\nclient-side, and send the whole thing back. Omitting `answers` entirely stores `{}`\nand wipes what was there.\n\n`timezone` is the exception: omit it or send `null` to keep the zone the user already\nhas (or `UTC` on a first write). When present it must be an IANA zone name; it is\nvalidated against the system tz database before anything is saved, so a bad value\nchanges nothing.\n\nTypically called right after a login that returned `is_new_user: true`, and again\nwhenever the user revisits their settings. Later plan generations pick the new values\nup automatically; plans already generated are not rewritten.",
        "operationId": "update_profile_v1_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The profile as stored, including the resolved timezone.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileView"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized` — missing, invalid or expired bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "`invalid_input` — `answers` is not a JSON object, one of its keys is not a string, or `timezone` is not a name the IANA database knows (e.g. `Asia/Taipei`, not `GMT+8`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` — too many requests for this user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerJWT": []
          }
        ]
      }
    },
...
    "/v1/plan-sessions": {
      "post": {
        "tags": [
          "plan-sessions"
        ],
        "summary": "Start plan generation from a goal",
        "description": "Where plan generation begins. **Only `goal` is required**; everything else is context.\n\n- `intake` — free-form answers about schedule, level and constraints. Whatever is missing\n  is what the service will ask about in the follow-up round.\n- `import_ids` — ids from `GET /v1/imports`. Each must belong to the caller and already be\n  `parsed`; anything else is rejected with 422 rather than silently ignored.\n- `trait_role_model_id` / `persona_role_model_id` — optional picks from `/v1/role-models`.\n- A live Google Calendar connection is picked up automatically, so the generated schedule\n  avoids the user's existing events. No field controls this.\n\nThe response is `202`, never a plan: generation runs on the queue. Poll\n`GET /v1/plan-sessions/{session_id}` and drive the UI off `status`:\n\n| status | what the client does |\n| --- | --- |\n| `collecting` | keep polling — context is being gathered |\n| `evaluating` | keep polling — the service is deciding whether it knows enough |\n| `questioning` | render `questions`, answer via `POST /v1/plan-sessions/{id}/answers` |\n| `generating` | keep polling — the three plans are being written |\n| `done` | `plans` holds three difficulty variants; let the user pick one |\n| `failed` | show `error`; the session is terminal and cannot be resumed |\n\n`GET /v1/jobs/{job_id}` reports the queue's own view of the same work, but the session row\nis the source of truth — a job record can expire while the session stays readable.",
        "operationId": "create_plan_session_v1_plan_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlanSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted. `session_id` is the resource to poll; `job_id` identifies the queued `plan.generate` job.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSessionResult"
                }
              }
            }
          },
          "422": {
            "description": "`invalid_input` — `goal` is empty or whitespace only, one of `import_ids` does not belong to the caller, or an import has not finished parsing (its `status` is not `parsed`). Body validation failures use the same code.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized` — the `Authorization: Bearer <jwt>` header is missing or malformed, or the token has expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` — too many requests from this caller in the last minute. The response carries `Retry-After`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerJWT": []
          }
        ]
      }
    },
    "/v1/plan-sessions/{session_id}": {
      "get": {
        "tags": [
          "plan-sessions"
        ],
        "summary": "Poll a plan session for questions or the finished plans",
        "description": "The one endpoint to poll after `POST /v1/plan-sessions` or a round of answers.\n\nWhich fields are populated depends entirely on `status`:\n\n- `questions` is non-empty **only** while `status` is `questioning`. Each question carries\n  its `options`, and whether a free-text answer (`allow_custom`) or a skip (`allow_skip`) is\n  accepted. `round` counts the follow-up rounds already asked; the service asks at most a\n  small fixed number and then generates with what it has.\n- `plans` is filled **only** once `status` is `done`, with the three difficulty variants —\n  each summarised with its weekly load (`sessions_per_week`, `total_minutes_per_week`) and\n  dates. All three start as `draft`; the user picks one by activating it with\n  `PATCH /v1/plans/{plan_id}` (`{\"status\": \"active\"}`).\n- `error` is set only when `status` is `failed`, which is terminal — start a new session.\n\nPolling is cheap but counts against the rate limit; a few seconds between calls is plenty.",
        "operationId": "get_plan_session_v1_plan_sessions__session_id__get",
        "security": [
          {
            "BearerJWT": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The session's current `status` and `round`, plus `questions` while `questioning` and `plans` once `done`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanSessionView"
                }
              }
            }
          },
          "404": {
            "description": "`not_found` — no session with this id belongs to the caller. Another user's session is indistinguishable from one that never existed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized` — the `Authorization: Bearer <jwt>` header is missing or malformed, or the token has expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` — too many requests from this caller in the last minute. The response carries `Retry-After`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/role-models": {
      "get": {
        "tags": [
          "role-models"
        ],
        "summary": "Browse the role model catalogue",
        "description": "The picker behind `trait_role_model_id` / `persona_role_model_id` on a plan session.\n\nRole models are the shared, team-curated library the plan generator borrows from: a *trait*\nmodel shapes how a plan is structured, a *persona* model shapes how it speaks. Filter with\n`kind`, and with `tags` repeated once per tag (`?tags=fitness&tags=beginner`) — `match=any`\nreturns anything carrying one of them, `match=all` only what carries every one.\n\n**Auth: the user's JWT.** Reading the catalogue is open to any signed-in user; only the\n`POST` / `PUT` / `DELETE` routes below are team-only.\n\nProxied straight from the Role Model Service, so the body and status are that service's,\nnot this one's. The error envelope is identical in both.",
        "operationId": "list_role_models_v1_role_models_get",
        "security": [
          {
            "BearerJWT": []
          }
        ],
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "default": [],
              "title": "Tags"
            }
          },
          {
            "name": "match",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "any",
                "all"
              ],
              "type": "string",
              "default": "any",
              "title": "Match"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A JSON array of role model summaries (`id`, `kind`, `name`, `tags`), proxied verbatim from the Role Model Service.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "`unauthorized` — the `Authorization: Bearer <jwt>` header is missing or malformed, or the token has expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "`invalid_input` — `limit` is outside 1–200, or `match` is not `any` / `all`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` — too many requests from this caller in the last minute. The response carries `Retry-After`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "role-models"
        ],
        "summary": "Create a role model (team only, X-API-Key)",
        "description": "Add an entry to the shared catalogue. **Content tooling only — not an app endpoint.**\n\n**Auth: `X-API-Key`, not a JWT.** The header is forwarded to the Role Model Service, which\nchecks it; no user token is involved and no user context is recorded. The catalogue is global,\nso anything created here is visible to every signed-in user.\n\nThe body is passed through unvalidated by this service: `kind` and a non-empty `name` are\nrequired, with optional `tags` (list of strings) and `content` (a free-form object whose shape\nthe Role Model Service defines). The status and body of the response are the upstream ones.",
        "operationId": "create_role_model_v1_role_models_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "`201` with the created record, proxied from the Role Model Service.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "`unauthorized` — the `X-API-Key` header is missing or does not match the Role Model Service's key. This is a team credential, not the user's JWT: a signed-in end user cannot reach this endpoint.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "`invalid_input` — the body is missing `kind` or `name`, or `name` is empty.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` — too many requests from this caller in the last minute. The response carries `Retry-After`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/role-models/tags": {
      "get": {
        "tags": [
          "role-models"
        ],
        "summary": "List the tags available for filtering",
        "description": "The vocabulary for the `tags` filter of `GET /v1/role-models`.\n\nUse it to build the filter UI instead of hard-coding tag names: the catalogue is curated by\nthe team and its tags change without an app release. Requires the user's JWT.",
        "operationId": "list_tags_v1_role_models_tags_get",
        "responses": {
          "200": {
            "description": "An object keyed by role model `kind`, each holding the tags in use for that kind.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "`unauthorized` — the `Authorization: Bearer <jwt>` header is missing or malformed, or the token has expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` — too many requests from this caller in the last minute. The response carries `Retry-After`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerJWT": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateSessionResult": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "job_id"
        ],
        "title": "CreateSessionResult",
        "description": "What both `POST /plan-sessions` and `POST /plan-sessions/{id}/answers` return."
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProfileView": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "answers": {
            "additionalProperties": true,
            "type": "object",
            "title": "Answers"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "answers",
          "timezone",
          "updated_at"
        ],
        "title": "ProfileView"
      },
...
      "UpdateProfileRequest": {
        "properties": {
          "answers": {
            "additionalProperties": true,
            "type": "object",
            "title": "Answers",
            "default": {}
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          }
        },
        "type": "object",
        "title": "UpdateProfileRequest"
      },
...
    },
    "securitySchemes": {
      "BearerJWT": {
        "type": "http",
        "description": "A JWT issued by `POST /v1/auth/google`. Send it as `Authorization: Bearer <token>`.",
        "scheme": "bearer"
      }
    }
  },
...
  "servers": [
...
}